Updates passed statuses to notification clicks

+++ library/src/main/java/com/novoda/downloadmanager/lib/NotificationDisplayer.java
@@ -99,7 +99,7 @@ class NotificationDisplayer {
             Uri uri = new Uri.Builder().scheme("active-dl").appendPath(tag).build();
             Intent clickIntent = new Intent(Constants.ACTION_LIST, uri, context, DownloadReceiver.class);
             clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_IDS, new long[]{batchId});
-            clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_STATUSES, new long[]{batchStatus});
+            clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_STATUSES, new int[]{batchStatus});
             builder.setContentIntent(PendingIntent.getBroadcast(context, 0, clickIntent, PendingIntent.FLAG_UPDATE_CURRENT));
             builder.setOngoing(true);
 
@@ -121,7 +121,7 @@ class NotificationDisplayer {
 
                 Intent clickIntent = new Intent(action, uri, context, DownloadReceiver.class);
                 clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_IDS, new long[]{batchId});
-                clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_STATUSES, new long[]{batchStatus});
+                clickIntent.putExtra(DownloadManager.EXTRA_NOTIFICATION_CLICK_DOWNLOAD_STATUSES, new int[]{batchStatus});
                 clickIntent.putExtra(DownloadReceiver.EXTRA_BATCH_ID, batchId);
                 builder.setContentIntent(PendingIntent.getBroadcast(context, 0, clickIntent, PendingIntent.FLAG_UPDATE_CURRENT));
             }
